Revert "weston evdev: Filter double press, double release"
This reverts commit 06d19df9899745d0fe0ec2920deccf7d4b2d1e3e.

Jonas Ådahl has explained why this is not the correct fix per linux
documentation. This is a quote from Jonas:

"It seems that what is done is avoid situations when a drop
would result in weird state transitions, but according to
documentation[0] what should be done is abort and ignore the
transaction and synchronize the appropriate states using ioctls. See
Documentation/input/event-codes.txt in the Linux kernel tarball:

* SYN_DROPPED:
  - Used to indicate buffer overrun in the evdev client's event queue.
    Client should ignore all events up to and including next SYN_REPORT
    event and query the device (using EVIOCG* ioctls) to obtain its
    current state."