shell: Only save position on fullscreen when not maximized
Fullscreen and maximize states require saving the size and position to know
how to restore properly. If a surface was already maximized when fullscreening,
don't save the position since maximize already has. The same logic might be
applied in reverse for going from fullscreen to maximize, but currently there
is no way to do this.