wobbly: Fix abrupt animation stop on ungrab
If the model animation ends while a grab is held, it doesn't
get a grab notification when resuming movement. This is fine
but releasing the grab during animation subsequently will
immediately end the model animation because it is in synced
state. Fix this by updating the synced state when a move
notifcation is received while grabbed.