wobbly: Fix resizing for wobbly surfaces
This problem was due to the fact that rotate changed the
surface position when rotating to offset the fact that
resize didn't keep the position updated as expected, that
is, position.p + 0.5 * size.p = center.p. Wobbly must also
make this assumption to sync correctly but the position
must be adjusted during resize, not on rotate motion.

As a side note, shell_surface_adjust_for_transform should
end up in the rotate plugin resize handler, if/when it is
ported to a standalone version.