gl-renderer: Workaround for opaque region issue
In weston, clients can set an opaque region which defines an
area where a black background will be drawn behind a surface.
This is a special case for RGBA textures with bad alpha data.
Wobbly uses an algorithm to define vertices that makes it
difficult to follow this hack. For now, just use the solid
shader if the opaque region is the same as the blend region
and the normal shader otherwise.